Fix various typos, man page issues and improve the documentation (#415)
This commit makes various different improvements to the documentation:
- sh.1: Backported (with changes) mandoc warning fixes from ksh2020
  for the ksh93(1) man page: <https://github.com/att/ast/pull/1406>
- Removed unnecessary spaces at the end of lines to fix a few other
  mandoc warnings.
- Fixed various typos and capitalization errors in the documentation.
- ANNOUNCE: Document the addition of the ${.sh.pid} variable
  (re: 9de65210).
- libast/man/str*: Update the man pages for the libast str* functions
  to improve how accurately each function is described.
- ksh93/README: Update regression test/compatibility notes to include
  OpenBSD 7.0, FreeBSD 13.0 and WSL running Ubuntu 20.04.
- Change a few places to store the return value from strlen in a
  size_t variable rather than signed int.
- comp/setlocale.c: To avoid confusion of two separate variables named
  lang, the function local variable has been renamed to langidx.

+++ src/cmd/ksh93/bltins/typeset.c
@@ -1556,7 +1556,7 @@ static void	print_attribute(register Namval_t *np,void *data)
 
 /*
  * print the nodes in tree <root> which have attributes <flag> set
- * of <option> is non-zero, no subscript or value is printed.
+ * if <option> is non-zero, no subscript or value is printed
  */
 
 static void print_scan(Sfio_t *file, int flag, Dt_t *root, int option,struct tdata *tp)
@@ -1566,7 +1566,7 @@ static void print_scan(Sfio_t *file, int flag, Dt_t *root, int option,struct tda
 	register int namec;
 	Namval_t *onp = 0;
 	char	*name=0;
-	int	len;
+	size_t	len;
 	sh.last_table=0;
 	flag &= ~NV_ASSIGN;
 	tp->scanmask = flag&~NV_NOSCOPE;
